app/backend/requirements.txt
alembic==1.16.2
black==23.7.0
bleach==6.2.0
blinker==1.9.0
click==8.2.1
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
flake8==6.1.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.12
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
greenlet==3.2.3
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.4.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mypy_extensions==1.1.0
mysqlclient==2.2.0
ordered-set==4.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==7.4.0
python-dotenv==1.0.0
python-magic==0.4.27
rich==13.9.4
SQLAlchemy==2.0.41
typing_extensions==4.14.0
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
flask-jwt-extended
